Report form is now a component
El formulario de reportado es un componente que muestra en el centro de la pantalla  unos checks con diferentes razones para reportar. Hay que pasarle 3 propiedades, la primera es la función que usan para mostrar y esconder el formulario, la segunda es el tipo de contenido que se está reportando, y la tercera es el valor de la primera función el cual deberá ser el Id del contenido a reportar
